Expanded modules to one class per file
