+++ GEECS-Scanner-GUI/geecs_scanner/data_acquisition/__init__.py
@@ -1,4 +1,8 @@
-from .data_acquisition import DeviceManager, DataLogger, ActionManager, SoundPlayer
+from .sound_player import SoundPlayer
+from .action_manager import ActionManager
+from .device_manager import DeviceManager
+from .data_logger import DataLogger
 from .database_dict_lookup import DatabaseDictLookup
+from .scan_data_manager import ScanDataManager
 from .scan_manager import ScanManager
 from .utils import ConsoleLogger, get_full_config_path, visa_config_generator
